+++ src/lib/components/common.ts
@@ -27,3 +27,13 @@ export function getDomain(domain) {
 export function generateRemoteEngine(destination) {
 	return `ssh://${destination.user}@${destination.ipAddress}:${destination.port}`;
 }
+
+export function dashify(str: string, options?: any): string {
+	if (typeof str !== 'string') return str;
+	return str
+		.trim()
+		.replace(/\W/g, (m) => (/[À-ž]/.test(m) ? m : '-'))
+		.replace(/^-+|-+$/g, '')
+		.replace(/-{2,}/g, (m) => (options && options.condense ? '-' : m))
+		.toLowerCase();
+}
